Remove executor interface
